Extract structured reporting export helpers into a partial
Second endpoint-decomposition slice. Moves the Workstation Structured Reporting
Export helper cluster out of the WorkstationEndpoints.cs core into a new sibling
partial WorkstationEndpoints.StructuredReportingExport.cs.

Moved (whole private static methods, closed call graph — each referenced only
from the inline WorkstationReportingStructuredExport endpoint lambda in
MapWorkstationEndpoints and from within the cluster):
ApplyWorkstationStructuredExportAuditHeaders,
IsWorkstationStructured{Csv,Json,Xlsx}Request,
BuildWorkstationStructuredExport{Csv,Xlsx,MetadataRows,DataDictionaryRows,ValidationRows,RowLineageRows},
AppendWorkstationStructuredExportCsvRow, EscapeWorkstationStructuredExportCsvValue.

The inline export endpoint lambda and the WorkstationStructuredXlsxContentType
const stay in core. Because WorkstationEndpoints is a single partial class this
is behavior-preserving: the lambda still calls the moved helpers across the
partial with no signature/route/DI change. Core drops 7680 -> 7488 lines; the
ratchet baseline is retightened to the new core cap.
